+++ python/src/internal.rs
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-use crate::utils::convert_vec_to_slice;
+use hudi::util::convert_vec_to_slice;
 use std::collections::HashMap;
 use std::path::PathBuf;
 use std::sync::OnceLock;
@@ -26,11 +26,11 @@ use arrow::pyarrow::ToPyArrow;
 use pyo3::{pyclass, pyfunction, pymethods, PyErr, PyObject, PyResult, Python};
 use tokio::runtime::Runtime;
 
-use crate::utils::vec_to_slice;
 use hudi::file_group::reader::FileGroupReader;
 use hudi::file_group::FileSlice;
 use hudi::table::builder::TableBuilder;
 use hudi::table::Table;
+use hudi::util::vec_to_slice;
 
 #[cfg(not(tarpaulin))]
 #[derive(Clone, Debug)]
